refactor(#1906): satisfy structural ratchets by extracting recall-concurrency helpers
The recall single-flight/concurrency work (#1906, PR #2015) grew several
grandfathered files past their #1520/#1995 ceilings. Extract the additions
into sibling modules and route the single-flight gate through the shared
access-setup capability projection so no baseline ceiling is raised and no
new scattered `config.*Enabled` read is introduced. Behavior and public APIs
are unchanged.

- access-recall-concurrency.ts: new module owning the per-principal semaphore
  + single-flight coalescing (acquireRecallSlot/withRecallConcurrency/raceAbort/
  attachFlightAbort/consumeFlight/followRecallFlight/runRecallDirect/
  createAndStartFlight/leadRecallFlight/coordinateRecall) as free functions
  over a RecallCoordinatorHost; access-service.recall() now delegates. Drops
  access-service.ts 5911->5884.
- access-recall-response.ts: new module holding assembleRecallResponse; drops
  access-recall-surface.ts 1650->1366.
- recall-concurrency-config.ts: new module parsing the concurrency knobs; keeps
  config.ts and types.ts at their ceilings.
- recall-timings.foldQueueWaitTiming: new helper folding the additive queue-wait
  phase; keeps recall-internal.ts at its ceiling.
- capabilities.resolveAccessSetupCapabilities gains recallSingleFlight so the
  gate is read only in the ratchet-exempt capabilities.ts (#1523).
- Tests updated to exercise the extracted free functions with identical
  invariants (concurrency, cancellation, capability parity).
